ENG-0000 feat(repo): Update data-populator main page UI (#902)
## Affected Packages

Apps

- [X] data populator
- [ ] portal
- [ ] template

Packages

- [X] 1ui
- [ ] api
- [ ] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview
This PR low-lift polishes up main page of data-populator. It includes a
refactor of tabs so all user forms are encapsulated into a tab card,
separated from the table so it can always be editted, updated.

I refrained from updating any business logic. The only business logic
change is how loadCSV is triggered when a new CSV is uploaded.

* Refactor tabs
* Restructure & polish UI

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [X] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)