ENG-4664 feat(portal,1ui): update queries in stake modal (#893)
## Affected Packages

Apps

- [ ] data populator
- [x] portal
- [ ] template

Packages

- [x] 1ui
- [ ] api
- [ ] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Based off branch/PR 4383 since there were some important updates to the
stake modal that had to be accounted for. Cannot merge until we get that
PR merged in.

- Swapped out vaultDetails fetcher for react query hook
- honestly can't remember some of the other updates since so much is
crammed into this due it being based off another branch. Once that is
merged and this is rebased, it will clear that all up.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)

---------

Co-authored-by: Jonathan Prozzi <jonathanprozzi@gmail.com>