ENG-4480 feat(portal,graphql): query consumer global activity feed (#947)
- Updates the global activity feed to use GraphQL and we also can now
support showing the user's positions on the atoms/triples in the feed.
This takes in an array of addresses.
- Modifies the events queries to support this feature and runs codegen
- Pagination works in the query, and added in a placeholder component --
doing pagination in a future PR
- Begins adding in new implementations of our utils in `misc.tsx` and
suffixes these with `New` -- same with the new component
implementations. This is to support the legacy versions until we fully
migrate. This is needed to ensure that everything builds.
- Note: We need to adjust the stake modals to work with the new types
without breaking the other implementations until we fully migrate